refactor : 코딩 컨벤션 점검
