+++ user-api/src/docs/asciidoc/index.adoc
@@ -241,7 +241,7 @@ include::{snippets}/user-controller-test/success-read-profile-test/response-fiel
 .error response
 |===
 |status|ErrorCode|message
-|400|USER_NOT_FOUND|해당 유저가 없습니다.
+|400|NOT_REGISTERED_MEMBER|해당 유저가 없습니다.
 |400|CANNOT_FOLLOW_YOURSELF|자기 자신을 팔로우할 수 없습니다."
 |===
 
@@ -275,7 +275,7 @@ include::{snippets}/user-controller-test/success-read-followings-test/response-f
 .error response
 |===
 |status|ErrorCode|message
-|400|USER_NOT_FOUND|해당 유저가 없습니다.
+|400|NOT_REGISTERED_MEMBER|해당 유저가 없습니다.
 |===
 
 === 5. 유저의 팔로워 리스트 조회
@@ -297,7 +297,7 @@ include::{snippets}/user-controller-test/success-read-followers-test/response-fi
 .error response
 |===
 |status|ErrorCode|message
-|400|USER_NOT_FOUND|해당 유저가 없습니다.
+|400|NOT_REGISTERED_MEMBER|해당 유저가 없습니다.
 |===
 
 === 6. 나의 팔로우 리스트 조회
