+++ user-api/src/main/java/zerobase/bud/chat/controller/ChatController.java
@@ -7,8 +7,7 @@ import org.springframework.messaging.simp.SimpMessagingTemplate;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import zerobase.bud.chat.dto.ChatDto;
-import zerobase.bud.chat.dto.ChatImageRequest;
-import zerobase.bud.chat.dto.ChatMessageRequest;
+import zerobase.bud.chat.dto.ChatMessage;
 import zerobase.bud.chat.service.ChatService;
 
 import javax.validation.Valid;
@@ -23,7 +22,7 @@ public class ChatController {
     private final SimpMessagingTemplate messagingTemplate;
 
     @MessageMapping("/message")
-    public void chatting(@RequestBody @Valid ChatMessageRequest request) {
+    public void chatting(@RequestBody @Valid ChatMessage.Request request) {
         ChatDto chat = chatService.chatting(
                 request.getMessage(), request.getChatroomId(), request.getSenderId());
         messagingTemplate.convertAndSend("/chatrooms/" + request.getChatroomId(), chat);
