+++ user-api/src/main/java/zerobase/bud/chat/dto/ChatImage.java
@@ -1,0 +1,24 @@
+package zerobase.bud.chat.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.NotNull;
+
+public class ChatImage {
+    @Getter
+    @NoArgsConstructor
+    @AllArgsConstructor
+    @Builder
+    public static class Request {
+        @NotNull
+        private Long senderId;
+        @NotNull
+        private Long chatroomId;
+        @NotBlank
+        private String imageCode;
+    }
+}
