--- user-api/src/main/java/zerobase/bud/chat/dto/ChatMessageRequest.java
@@ -1,22 +1,0 @@
-package zerobase.bud.chat.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class ChatMessageRequest {
-    @NotNull
-    private Long senderId;
-    @NotNull
-    private Long chatroomId;
-    @NotBlank
-    private String message;
-}
