+++ user-api/src/main/java/zerobase/bud/common/type/ErrorCode.java
@@ -10,7 +10,6 @@ public enum ErrorCode {
     NOT_FOUND_POST("존재하지 않는 게시물입니다."),
     INVALID_INITIAL_VALUE("잘못된 초기값입니다."),
     INVALID_TOTAL_COMMIT_COUNT("유효하지 않은 총 커밋 수 입니다."),
-    USER_NOT_FOUND("해당 유저가 없습니다."),
     CANNOT_FOLLOW_YOURSELF("자기 자신을 팔로우할 수 없습니다."),
     WEB_SOCKET_ERROR("웹 소켓 연결에 오류가 발생했습니다."),
     CHATROOM_NOT_FOUND("해당 채팅방 정보가 없습니다."),
