--- user-api/src/main/java/zerobase/bud/common/util/Constant.java
@@ -1,14 +1,0 @@
-package zerobase.bud.common.util;
-
-public abstract class Constant {
-    public static final int CHATROOM_SIZE_PER_PAGE = 5;
-    public static final int CHAT_SIZE_PER_PAGE = 10;
-
-    public static final String TOKEN_PREFIX = "Bearer ";
-
-    public static final String MAXIMUM_LEVEL_CODE = "찬란한_나무";
-
-    public static final String SESSION = "CHATROOM_SESSION";
-
-    public static final String CHATROOM = "CHATROOM";
-}
