+++ user-api/src/main/java/zerobase/bud/common/util/Constants.java
@@ -1,6 +1,17 @@
-package zerobase.bud.news.type;
+package zerobase.bud.common.util;
 
-public final class Constants {
+public abstract class Constants {
+    public static final int CHATROOM_SIZE_PER_PAGE = 5;
+    public static final int CHAT_SIZE_PER_PAGE = 10;
+
+    public static final String TOKEN_PREFIX = "Bearer ";
+
+    public static final String MAXIMUM_LEVEL_CODE = "찬란한_나무";
+
+    public static final String SESSION = "CHATROOM_SESSION";
+
+    public static final String CHATROOM = "CHATROOM";
+
     public static final String[] NAVER_NEWS_API_KEYWORDS = {
             "자바", "자바스크립트", "파이썬", "알고리즘",
             "코딩테스트", "개발", "개발자", "인공지능",
@@ -9,21 +20,30 @@ public final class Constants {
             "C언어", "코딩", "프로그래밍"
     };
     public static final String[] NAVER_NEWS_API_SORT = {"date", "sim"};
+
     public static final int RESULT_COUNT = 50;
 
     public static final String NAVER_NEWS_API_BASIC_URL = "https://openapi.naver.com/v1/search/news.json?";
+
     public static final String NAVER_NEWS_ITEMS = "items";
 
     public static final String NAVER_CLIENT_ID_KEY = "X-Naver-Client-Id";
+
     public static final String NAVER_CLIENT_SECRET_KEY = "X-Naver-Client-Secret";
 
     public static final String PARSING_POSSIBLE_DOMAIN = "n.news.naver.com";
 
     public static final String CONTENT_SELECTOR = "div#ct";
+
     public static final String ARTICLE_SELECTOR = "#newsct_article";
+
     public static final String COMPANY_TITLE_ATTR_KEY = "title";
+
     public static final String COMPANY_LOGO_IMG_SELECTOR = "img.media_end_head_top_logo_img";
+
     public static final String FIRST_IMG_SELECTOR = "#img1";
+
     public static final String FIRST_IMG_ATTR_KEY = "data-src";
+
     public static final String JOURNALIST_SELECTOR = "span.byline_s";
 }
