+++ user-api/src/main/java/zerobase/bud/oauth/service/CustomOAuth2UserService.java
@@ -11,7 +11,7 @@ import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.stereotype.Service;
 import zerobase.bud.domain.Member;
 import zerobase.bud.repository.MemberRepository;
-import zerobase.bud.oauth.attribute.OAuthAttribute;
+import zerobase.bud.oauth.dto.OAuthAttribute;
 
 import java.util.Collections;
 import java.util.Optional;
